refactor(group): UseCase 파일 기능별 폴더 구조로 재조직
- 타이머, 출석부, 그룹관리, 검색, 채팅 기능별 폴더 생성

- 기존 UseCase 파일들을 관련 기능 폴더로 이동

- 코드 내용 변경 없이 구조만 정리
