+++ lib/group/domain/usecase/management/get_group_detail_use_case.dart
