+++ lib/group/domain/usecase/management/mock_get_group_detail_use_case.dart
