+++ lib/group/presentation/group_attendance/attendance_notifier.dart
@@ -1,9 +1,9 @@
 import 'package:devlink_mobile_app/core/utils/app_logger.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/get_attendance_by_month_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/attendance/get_attendance_by_month_use_case.dart';
 import 'package:devlink_mobile_app/group/module/group_di.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 import 'package:intl/date_symbol_data_local.dart';
+import 'package:intl/intl.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import 'attendance_action.dart';
