+++ lib/group/presentation/group_detail/group_detail_notifier.dart
@@ -8,10 +8,10 @@ import 'package:devlink_mobile_app/core/utils/app_logger.dart';
 import 'package:devlink_mobile_app/core/utils/time_formatter.dart';
 import 'package:devlink_mobile_app/group/domain/model/group_member.dart';
 import 'package:devlink_mobile_app/group/domain/model/timer_activity_type.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/get_group_detail_use_case.dart';
 import 'package:devlink_mobile_app/group/domain/usecase/get_group_members_use_case.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/record_timer_activity_use_case.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/stream_group_member_timer_status_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/management/get_group_detail_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/timer/record_timer_activity_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/timer/stream_group_member_timer_status_use_case.dart';
 import 'package:devlink_mobile_app/group/module/group_di.dart';
 import 'package:devlink_mobile_app/group/presentation/group_detail/group_detail_action.dart';
 import 'package:devlink_mobile_app/group/presentation/group_detail/group_detail_state.dart';
