+++ lib/group/presentation/group_search/group_search_notifier.dart
@@ -2,8 +2,8 @@ import 'package:devlink_mobile_app/core/auth/auth_provider.dart';
 import 'package:devlink_mobile_app/core/service/search_history_service.dart';
 import 'package:devlink_mobile_app/core/utils/app_logger.dart';
 import 'package:devlink_mobile_app/group/domain/model/group.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/join_group_use_case.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/search_groups_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/management/join_group_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/search/search_groups_use_case.dart';
 import 'package:devlink_mobile_app/group/module/group_di.dart';
 import 'package:devlink_mobile_app/group/presentation/group_search/group_search_action.dart';
 import 'package:devlink_mobile_app/group/presentation/group_search/group_search_state.dart';
