+++ lib/group/presentation/group_setting/group_settings_notifier.dart
@@ -5,10 +5,10 @@ import 'package:devlink_mobile_app/core/utils/app_logger.dart';
 import 'package:devlink_mobile_app/core/utils/image_compression.dart';
 import 'package:devlink_mobile_app/group/domain/model/group.dart';
 import 'package:devlink_mobile_app/group/domain/model/group_member.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/get_group_detail_use_case.dart';
 import 'package:devlink_mobile_app/group/domain/usecase/get_group_members_use_case.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/leave_group_use_case.dart';
-import 'package:devlink_mobile_app/group/domain/usecase/update_group_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/management/get_group_detail_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/management/leave_group_use_case.dart';
+import 'package:devlink_mobile_app/group/domain/usecase/management/update_group_use_case.dart';
 import 'package:devlink_mobile_app/group/module/group_di.dart';
 import 'package:devlink_mobile_app/group/presentation/group_setting/group_settings_action.dart';
 import 'package:devlink_mobile_app/group/presentation/group_setting/group_settings_state.dart';
